chore: merge all fuzzers and clean the cmake
(cherry picked from commit 432393500065238d851480f7a6c9b072293c91d5)
