+++ .clusterfuzzlite/build.sh
@@ -3,7 +3,7 @@
 # build fuzzers
 
 pushd tests/fuzzing
-cmake -DBOLOS_SDK=../../BOLOS_SDK -Bbuild -H.
-make -C build
-mv ./build/fuzz_app_eth "${OUT}"
+cmake -DBOLOS_SDK=$(pwd)/../../BOLOS_SDK -B build -S .
+cmake --build build
+mv ./build/fuzzer "${OUT}"
 popd
